// pslpwcl/src/app/modules/pslpapi/model/incServizi.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Decodifica } from './decodifica';
import { InTipoContatto } from './inTipoContatto';
import { IncServCateg } from './incServCateg';
import { MsgCanale } from './msgCanale';
import { StepFlusso } from './stepFlusso';
import { TipoOperat } from './tipoOperat';

export interface IncServizi { 
    codAmbito?: string;
    codTipoAgenda?: string;
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dFine?: Date;
    dInizio?: Date;
    dInserim?: Date;
    decodificaTipoAgenda?: Decodifica;
    descrEstSilwebTIncServizi?: string;
    descrSilwebTIncServizi?: string;
    descrTestoAggDisdetta?: string;
    descrTestoAggNoPresente?: string;
    descrTestoAggPrenota?: string;
    descrTestoAggPromemoria?: string;
    descrTestoAnnulla?: string;
    descrTestoSposta?: string;
    durataMinutiIncontro?: number;
    flgBlocca?: string;
    flgInvioConferma?: string;
    flgInvioPromemoria?: string;
    flgPrenotaCittadino?: string;
    flgPresenzaVincoli?: boolean;
    flgProvincia?: string;
    flgSposta?: string;
    idSilwebTIncServStandard?: number;
    idSilwebTIncServizi?: number;
    listaModalitaErogazione?: Array<InTipoContatto>;
    listaMsgCanale?: Array<MsgCanale>;
    listaStepFlusso?: Array<StepFlusso>;
    numGiorniMailPromemoria?: number;
    numGiorniMargine?: number;
    numGiorniPrenota?: number;
    numGiorniSposta?: number;
    numMaxPartecipantiAgende?: number;
    numMinPartecipantiAgende?: number;
    silwebTIncServCateg?: IncServCateg;
    silwebTTipoOperat?: TipoOperat;
    timestamp?: number;
}